# docs/swagger.yaml
definitions:
  github_com_Vovarama1992_emelya-go_internal_money_deposit_model.Deposit:
    properties:
      amount:
        type: number
      approved_at:
        type: string
      block_until:
        type: string
      created_at:
        type: string
      daily_reward:
        type: number
      id:
        type: integer
      status:
        $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_money_deposit_model.Status'
      user_id:
        type: integer
    type: object
  github_com_Vovarama1992_emelya-go_internal_money_deposit_model.Status:
    enum:
    - pending
    - approved
    - closed
    type: string
    x-enum-comments:
      StatusApproved: активный
      StatusClosed: закрыт, разблокирован
      StatusPending: ожидает подтверждения
    x-enum-varnames:
    - StatusPending
    - StatusApproved
    - StatusClosed
  github_com_Vovarama1992_emelya-go_internal_money_operation_model.Operations:
    properties:
      deposits:
        items:
          $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_money_deposit_model.Deposit'
        type: array
      rewards:
        items:
          $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_money_reward_model.Reward'
        type: array
      withdrawals:
        items:
          $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_money_withdrawal_model.Withdrawal'
        type: array
    type: object
  github_com_Vovarama1992_emelya-go_internal_money_reward_model.Reward:
    properties:
      amount:
        type: number
      created_at:
        type: string
      deposit_id:
        type: integer
      id:
        type: integer
      last_accrued_at:
        type: string
      type:
        $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_money_reward_model.RewardType'
      user_id:
        type: integer
      withdrawn:
        type: number
    type: object
  github_com_Vovarama1992_emelya-go_internal_money_reward_model.RewardType:
    enum:
    - deposit
    - referral
    type: string
    x-enum-varnames:
    - RewardTypeDeposit
    - RewardTypeReferral
...
  github_com_Vovarama1992_emelya-go_internal_money_withdrawal_model.Withdrawal:
    properties:
      amount:
        type: number
      approved_at:
        type: string
      created_at:
        type: string
      id:
        type: integer
      reason:
        type: string
      rejected_at:
        type: string
      reward_id:
        type: integer
      status:
        $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_money_withdrawal_model.WithdrawalStatus'
      user_id:
        type: integer
    type: object
  github_com_Vovarama1992_emelya-go_internal_money_withdrawal_model.WithdrawalStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - WithdrawalStatusPending
    - WithdrawalStatusApproved
    - WithdrawalStatusRejected
  github_com_Vovarama1992_emelya-go_internal_user_model.User:
    properties:
      cardNumber:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      isEmailVerified:
        type: boolean
      isPhoneVerified:
        type: boolean
      lastName:
        type: string
      login:
        type: string
      passwordHash:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      referrerID:
        type: integer
      role:
        $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_user_model.UserRole'
    type: object
  github_com_Vovarama1992_emelya-go_internal_user_model.UserRole:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
...
  internal_user_http.AddReferralRequest:
    properties:
      referrer_id:
        type: integer
      user_id:
        type: integer
    required:
    - referrer_id
    - user_id
    type: object
  internal_user_http.AdminUpdateProfileRequest:
    properties:
      card_number:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      user_id:
        type: integer
    required:
    - user_id
    type: object
  internal_user_http.UpdateProfileRequest:
    properties:
      card_number:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      patronymic:
        type: string
      phone:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/admin/user/add-referal:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID пользователя и ID реферала
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_user_http.AddReferralRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Админ добавляет реферала пользователю
      tags:
      - admin-user
  /api/admin/user/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_user_model.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить всех пользователей
      tags:
      - user
  /api/admin/user/operations:
    get:
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_money_operation_model.Operations'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 'Админ: все операции пользователя (депозиты, выводы, награды)'
      tags:
      - admin-user
  /api/admin/user/search-id:
    get:
      parameters:
      - description: ID пользователя
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_user_model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Найти пользователя по ID
      tags:
      - user
  /api/admin/user/update-profile:
    post:
      consumes:
      - application/json
      parameters:
      - description: Обновляемые поля пользователя
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_user_http.AdminUpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Админ обновляет профиль пользователя
      tags:
      - admin-user
...
  /api/user/update-profile:
    post:
      consumes:
      - application/json
      parameters:
      - description: Обновляемые поля
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_user_http.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить профиль (самостоятельно)
      tags:
      - user
...
swagger: "2.0"